changed size from int to long
git-svn-id: http://svn.littleshoot.org/svn/littleshoot/trunk/common/download@4648 8d7aff80-242d-0410-bcee-ed6f15385b3d

+++ src/main/java/org/lastbamboo/common/download/DummySha1Downloader.java
@@ -77,7 +77,7 @@ public final class DummySha1Downloader<DsT extends DownloaderState>
         return m_delegate.getIncompleteFile ();
         }
     
-    public int getSize ()
+    public long getSize ()
         {
         return m_delegate.getSize ();
         }
