+++ src/main/java/org/lastbamboo/common/download/Sha1Downloader.java
@@ -89,7 +89,7 @@ public final class Sha1Downloader<DsT extends DownloaderState>
         return m_delegate.getIncompleteFile ();
         }
     
-    public int getSize ()
+    public long getSize ()
         {
         return m_delegate.getSize ();
         }
